QA_Automation/requirements.txt
# Core testing framework
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.3.1
pytest-rerunfailures==12.0
pytest-json-report==1.5.0
pytest-cov==4.1.0

# API testing and HTTP client
requests==2.31.0
requests-oauthlib==1.3.1

# Data handling and validation
jsonschema==4.19.2
pydantic==2.5.0
pandas==2.1.3
openpyxl==3.1.2

# Configuration and environment management
python-decouple==3.8
pyyaml==6.0.1

# Logging and reporting
loguru==0.7.2
allure-pytest==2.13.2

# Faker for test data generation
faker==20.1.0

# Database support (optional)
sqlalchemy==2.0.23
pymongo==4.6.0

# Utilities
retry==0.9.2
deepdiff==6.7.1
jmespath==1.0.1

# Development tools
black==23.11.0
flake8==6.1.0
pre-commit==3.5.0 